% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bodymapRat.R
\docType{data}
\name{bodymapRat}
\alias{bodymapRat}
\title{Experimental dataset from the rat BodyMap project}
\format{A SummarizedExperiment object with 652 RNA-seq
samples (columns).}
\description{
This package contains a SummarizedExperiment
from the Yu et al. (2013) paper that performed the 
rat BodyMap across 11 organs and 4 developmental stages. 
Raw FASTQ files were downloaded and mapped using STAR.  
FlowSorted.DLPFC.450k data set from Bioconductor. 
The purpose is to create an example object for the man
pages and vignette in this package. 

The object was created using the script in /inst and
located in the /data folder.
}
